requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
ruff==0.6.6
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
